+++ Westwind.RazorHosting.Tests/StringHostTests.cs
@@ -21,22 +21,19 @@ namespace RazorHostingTests
             //
         }
 
-     
-
-      
 
         [TestMethod]
         public void BasicStringHostTest()
         {
             var host = new RazorStringHostContainer();
-            
+
             // add model assembly - ie. this assembly
             host.AddAssemblyFromType(this);
 
             host.UseAppDomain = false;
 
             host.Start();
-              
+
             Person person = new Person()
             {
                 Name = "Rick",
@@ -48,16 +45,16 @@ namespace RazorHostingTests
                     City = "Paia"
                 }
             };
-            
-            string result = host.RenderTemplate(Templates.BasicTemplateStringWithPersonModel,person);
-            
+
+            string result = host.RenderTemplate(Templates.BasicTemplateStringWithPersonModel, person);
+
             Console.WriteLine(result);
             Console.WriteLine("---");
             Console.WriteLine(host.Engine.LastGeneratedCode);
 
             if (result == null)
                 Assert.Fail(host.ErrorMessage);
-            
+
             host.Stop();
         }
 
@@ -91,7 +88,7 @@ namespace RazorHostingTests
                 }
             };
 
-            
+
             string template = @"
 <div>@Model.Name
 <div>
@@ -100,15 +97,15 @@ namespace RazorHostingTests
         <div>@addr.Street, @addr.Phone</div>    
 }
 </div>
-";   
+";
 
-            string result = host.RenderTemplate(template, person,inferModelType: true);
+            string result = host.RenderTemplate(template, person, inferModelType: true);
 
             Console.WriteLine(result);
-            Console.WriteLine("---");            
+            Console.WriteLine("---");
 
             Assert.IsNotNull(result, "Result shouldn't be null: " + host.ErrorMessage);
-                
+
             host.Stop();
         }
 
@@ -158,6 +155,71 @@ namespace RazorHostingTests
 
             host.Stop();
         }
- 
+
+
+
+
+        /// <summary>
+        /// Explicit template failure when a runtime error occurs
+        /// </summary>
+        [TestMethod]
+        public void BasicStringHostRuntimeErrorExceptionTest()
+        {
+            var host = new RazorStringHostContainer()
+            {
+                ThrowExceptions = true,
+                UseAppDomain = false
+            };
+
+            // add model assembly - ie. this assembly
+            host.AddAssemblyFromType(this);
+
+
+
+            host.Start();
+
+            Person person = new Person()
+            {
+                Name = "Rick",
+                Company = "West Wind",
+                Entered = DateTime.Now,
+                Address = new Address()
+                {
+                    Street = "32 Kaiea",
+                    City = "Paia"
+                }
+            };
+
+
+            string template = @"
+@{
+   Model.Name = null;   
+}
+<div>
+    Fail here with Null exception: 
+    @Model.Name.ToLower()
+<div>
+";
+            bool isException = false;
+            string result = null;
+            try
+            {
+                result = host.RenderTemplate(template, person, inferModelType: true);
+            }
+            catch (RazorHostContainerException ex)
+            {
+                isException = true;
+                Assert.IsNull(result, "Result should have failed with a runtime error.");
+                Console.WriteLine(ex.InnerException.Message);
+                Console.WriteLine(ex.InnerException.Source);
+                Console.WriteLine(ex.InnerException.StackTrace);
+                Console.WriteLine(ex.GeneratedSourceCode);
+
+            }
+            Console.WriteLine(result);
+            Console.WriteLine(host.ErrorMessage);
+
+            host.Stop();
+        }
     }
 }