+++ Westwind.RazorHosting/HostContainers/RazorFolderHostContainer.cs
@@ -297,8 +297,6 @@ namespace Westwind.RazorHosting
             return Path.GetFileNameWithoutExtension(fileName).Replace("\\", "_");
         }
 
-     
-     
     }
 
     /// <summary>