+++ ChangeLog.md
@@ -1,6 +1,13 @@
 # Westwind.RazorHosting Change Log
 
-## Version 3.1
+### Version 3.2
+*not released yet*
+
+* **Host Container ThrowExceptions Option**  
+You can now have the Razor Host Containers throw exceptions when an error occurs rather than returning error information on the error message property. A custom exception is thrown that provides more information about the current request and engine.
+
+
+### Version 3.1
 *May 29th, 2017*
 
 * **Support for Layout Pages with RazorFolderHostContainer**  
@@ -17,7 +24,7 @@ do not have this feature and must explicitly specify the model type.
 * **[Updated Documentation Site](https://west-wind.com/files/tools/razorhosting/docs/)**   
 Updated the documentation site to use the latest [Help Builder](https://helpuilder.west-wind.com) templates from old very outdated old MSDN style templates. Yikes! :smile: Updated content for version 3.0 improvements and added latest generated class documentation.
 
-## Version 3.0
+### Version 3.0
 *July 8th, 2014*
 
 * **Update to Razor Version 3**  
@@ -44,7 +51,7 @@ they are released. Default Dispose() behavior stops the engine and releases
 all cached assemblies.
 
 
-## Version 1.01
+### Version 1.01
 *December 20th, 2012*
 
 * **Changed default bin folder location**